ensured circular references among collections are rejected
assigning B as subcollection of A should fail if A is already a
subcollection of B

saving the respective collection still succeeds, but any circular
subcollection assignments are skipped and the user notified accordingly

NB: This cannot be implemented using validations since sub-collection
assignments occur in after_save, i.e. post-validation.