Implement a parser for an internal, canonical triple format.
The NTriples parser (and potentially others) are subclasses of that one.
The idea is to internally work with that triple format (as we do now,
but not explicitely so) using the centrally defined 'grammar', which
can be included anywhere as a module. This allows for more robust
string token handling in methods like build_from_rdf.