Implement inline_notes processing and Concept#destroy_later
Notes are now updated in the UI via Concep#inline_notes={...} like Labels.
Please note that this now requires _all_ notes to be passed in an update,
partial updates ("patch" as Rails uses it) are no longer supported. The
same is required for change notes!

Concept#destroy_later is used to remove an associated object from the
corresponding association collection in processing before_validation and
to actually destroy the object after_save.